cp-189: * rename chat related files
