+++ frontend/src/packages/chats/chats.ts
@@ -2,7 +2,7 @@ import { config } from '#libs/packages/config/config.js';
 import { http } from '#libs/packages/http/http.js';
 import { storage } from '#libs/packages/storage/storage.js';
 
-import { ChatApi } from './chat-api.js';
+import { ChatApi } from './chats-api.js';
 
 const chatApi = new ChatApi({
   baseUrl: config.ENV.API.ORIGIN_URL,
