--- frontend/src/slices/chat/chat.ts
@@ -1,1 +1,0 @@
-export { reducer } from './chat.slice.js';
