+++ frontend/src/slices/chats/chats.ts
@@ -1,0 +1,1 @@
+export { reducer } from './chats.slice.js';
